Change font size to a float (closes #1952)
This commit changes the 'pointSize' parameter in Font, Text
and Text3D from an int to a float, allowing e.g. 14.5pt text.
Note that when hinting is enabled, font metrics are snapped
to pixel boundaries, so the effect may be hard to see unless
you also set UI.FontHintLevel to LIGHT or NONE.

This is a change to the public API, but existing code (including
scripts) should compile and run as before.

+++ Source/Urho3D/UI/Font.h
@@ -65,7 +65,7 @@ public:
     void SetScaledGlyphOffset(const Vector2& offset);
 
     /// Return font face. Pack and render to a texture if not rendered yet. Return null on error.
-    FontFace* GetFace(int pointSize);
+    FontFace* GetFace(float pointSize);
 
     /// Return font type.
     FontType GetFontType() const { return fontType_; }
@@ -89,9 +89,9 @@ private:
     /// Load font glyph offset parameters from an optional XML file. Called internally when loading TrueType fonts.
     void LoadParameters();
     /// Return font face using FreeType. Called internally. Return null on error.
-    FontFace* GetFaceFreeType(int pointSize);
+    FontFace* GetFaceFreeType(float pointSize);
     /// Return bitmap font face. Called internally. Return null on error.
-    FontFace* GetFaceBitmap(int pointSize);
+    FontFace* GetFaceBitmap(float pointSize);
 
     /// Created faces.
     HashMap<int, SharedPtr<FontFace> > faces_;
